long long

+++ input.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main(){
-    int inputValue ;
-    scanf("%d", &inputValue);
-    printf("%d\n",inputValue);
+    long long  int inputValue ;
+    scanf("%lld", &inputValue);
+    printf("%lld\n",inputValue);
 }
